make view & viewModel & model groups
