+++ github_issues_list_app/Class/View/TopViewController.swift
@@ -8,17 +8,6 @@
 
 import UIKit
 
-struct CellData: Decodable {
-    let title: String
-    let state: String
-    let body: String
-    let html_url: String
-}
-
-//struct IssueResult: Decodable {
-//    let results: [CellData]
-//}
-
 class ViewController: UIViewController {
     @IBOutlet private weak var tableView: UITableView!
     let api = ItemRepository()
@@ -27,14 +16,12 @@ class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-//        print(self)
         tableView.delegate = self
         tableView.dataSource = self
-//
-        fetch()
+        loadCellData()
     }
 
-    func fetch() {
+    private func loadCellData() {
         api.fetch(completion: { [weak self] e in
             self?.tableDataSources = e
             self?.tableView.reloadData()
@@ -52,8 +39,6 @@ class ViewController: UIViewController {
 extension ViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         print("selected! \(indexPath.row)")
-//        tableView.deselectRow(at: indexPath, animated: true)
-//        performSegue(withIdentifier: "toNextViewController", sender: nil)
     }
 }
 
@@ -61,40 +46,11 @@ extension ViewController: UITableViewDelegate {
 extension ViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-//        print(tableDataSources[indexPath.row].title)
         cell.textLabel?.text = tableDataSources[indexPath.row].title
         return cell
     }
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-//        print(tableDataSources.count)
         return tableDataSources.count
-    }
-}
-
-import Alamofire
-
-//private let session: Session = {
-//    let manager = ServerTrustManager(evaluators: ["http://localhost:3000/": DisabledTrustEvaluator()])
-//    let configuration = URLSessionConfiguration.af.default
-//
-//    return Session(configuration: configuration, serverTrustManager: manager)
-//}()
-
-class ItemRepository {
-    func fetch(completion: @escaping ([CellData]) -> Void) {
-        let url: String = "http://localhost:3000/issues/index"
-        AF.request(url, method: .get, encoding: JSONEncoding.default).responseJSON { response in
-//            debugPrint(response)
-            switch response.result {
-            case .success:
-                guard let data = response.data else { return }
-                let issue = try? JSONDecoder().decode([CellData].self, from: data)
-                completion(issue!)
-            case .failure(let error):
-                print("error happened!!")
-                print(error)
-            }
-        }
     }
 }
